fix:Fix sort tasks when add new task or mark as completed or uncompleted or change position task by user on drag and drop

+++ src/server.php
@@ -9,17 +9,19 @@ $request_body = file_get_contents('php://input');
 $request = json_decode($request_body);
 if (isset($request->todo)) {
     $item = filter_var($request->todo, FILTER_SANITIZE_STRING);
-    $pos = filter_var($request->pos, FILTER_VALIDATE_INT);
-    if (!empty($item) AND !empty($pos)) {
-        $insert = mysqli_query($con, "INSERT INTO todo (item,todoStatu,sort) VALUES ('$item',1,$pos)");
-        if (isset($insert)) {
-            $query = mysqli_query($con, "SELECT * from todo WHERE todoStatu=1 ORDER BY sort ASC");
-            while ($data = mysqli_fetch_object($query)) {
-                $output["todos"][] = $data;
-            }
-            $query = mysqli_query($con, "SELECT * from todo WHERE todoStatu=2 ORDER BY sort ASC");
-            while ($data = mysqli_fetch_object($query)) {
-                $output["completed"][] = $data;
+    if (!empty($item)) {
+        $updatedSort = mysqli_query($con, "UPDATE todo SET sort = sort +1 WHERE todoStatu=1");
+        if (isset($updatedSort)) {
+            $insert = mysqli_query($con, "INSERT INTO todo (item,todoStatu,sort) VALUES ('$item',1,1)");
+            if (isset($insert)) {
+                $query = mysqli_query($con, "SELECT * from todo WHERE todoStatu=1 ORDER BY sort ASC");
+                while ($data = mysqli_fetch_object($query)) {
+                    $output["todos"][] = $data;
+                }
+                $query = mysqli_query($con, "SELECT * from todo WHERE todoStatu=2 ORDER BY sort ASC");
+                while ($data = mysqli_fetch_object($query)) {
+                    $output["completed"][] = $data;
+                }
             }
         } else {
             $output['msg'] = "error cannot execute query ";
@@ -28,19 +30,25 @@ if (isset($request->todo)) {
 
     }
 
-} elseif (isset($request->delete)) {
+} elseif (isset($request->delete) and isset($request->type) and isset($request->sort)) {
     $id = filter_var($request->delete, FILTER_VALIDATE_INT);
-    $deleted = mysqli_query($con, "DELETE FROM todo WHERE id = $id");
-    if ($deleted) {
-        $query = mysqli_query($con, "SELECT * from todo WHERE todoStatu=1 ORDER BY sort ASC");
-        while ($data = mysqli_fetch_object($query)) {
-            $output["todos"][] = $data;
-        }
-        $query = mysqli_query($con, "SELECT * from todo WHERE todoStatu=2 ORDER BY sort ASC");
-        while ($data = mysqli_fetch_object($query)) {
-            $output["completed"][] = $data;
+    $type = filter_var($request->type, FILTER_VALIDATE_INT);
+    $sort = filter_var($request->sort, FILTER_VALIDATE_INT);
+    $updateUncompletedSort = mysqli_query($con, "UPDATE todo SET sort = sort - 1 WHERE todoStatu = $type AND sort > $sort");
+    if (isset($updateUncompletedSort)) {
+        $deleted = mysqli_query($con, "DELETE FROM todo WHERE id = $id");
+        if ($deleted) {
+            $query = mysqli_query($con, "SELECT * from todo WHERE todoStatu=1 ORDER BY sort ASC");
+            while ($data = mysqli_fetch_object($query)) {
+                $output["todos"][] = $data;
+            }
+            $query = mysqli_query($con, "SELECT * from todo WHERE todoStatu=2 ORDER BY sort ASC");
+            while ($data = mysqli_fetch_object($query)) {
+                $output["completed"][] = $data;
+            }
         }
     }
+
 } elseif (isset($request->edit) and isset($request->item)) {
     $id = filter_var($request->edit, FILTER_VALIDATE_INT);
     $item = filter_var($request->item, FILTER_SANITIZE_STRING);
